fix(macos): Correctly get current display index (main)
The previous implementation used `[NSEvent mouseLocation]` and
`[NSScreen screens]` which are deprecated and don't work well
with multiple displays. It also had issues when displays had
overlapping frames.

This commit replaces it with `CGEventGetLocation`,
`CGGetActiveDisplayList` and `CGDisplayBounds` which are the
recommended way to get the current display index on macOS.  This
fixes the issue where the incorrect display index was returned
when using multiple displays.